Now need to specify numpy version in Python setup.
Something is strange in the QIIME build. Numpy version needs to be adjusted to a more recent version since QIIME has been build with that before. This specification was not necessary before. print-qiime-config.py does not seem to work but all the rest of the commands seems to be fine.